+++ CHANGELOG.md
@@ -6,8 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-## [v1.0.2](https://github.com/linna/dotenv/compare/v1.0.1...v1.0.2) - 2019-XX-XX
+## [v1.1.0](https://github.com/linna/dotenv/compare/v1.0.1...v1.1.0) - 2020-03-28
 
 ### Added
 * backslash in front of native functions